requirements.txt
alembic==1.7.5
api==0.0.7
asgiref==3.3.4
banal==1.0.6
certifi==2021.5.30
chardet==4.0.0
dataset==1.5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-cors-headers==3.7.0
django-import-export==2.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
et-xmlfile==1.1.0
greenlet==1.1.2
idna==2.10
importlib-metadata==4.8.2
importlib-resources==5.4.0
Mako==1.1.6
MarkupPy==1.14
MarkupSafe==2.0.1
nose==1.3.7
numpy==1.21.4
numpy-financial==1.0.0
odfpy==1.4.1
openpyxl==3.0.7
PyJWT==2.1.0
pytz==2021.1
PyYAML==6.0
requests==2.25.1
SQLAlchemy==1.4.27
sqlparse==0.4.1
tablib==3.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.6.0
